feat(Prize): 총 수익률을 계산하여 출력하는 기능 구현

+++ src/main/java/lotto/EarningRate.java
@@ -4,7 +4,7 @@ public class EarningRate {
 	public EarningRate() {
 	}
 
-	public static double calculate(int paidMoney, int earningMoney) {
+	public static double calculate(int paidMoney, Long earningMoney) {
 		return (double) earningMoney / paidMoney * 100;
 	}
 }
